requirements.txt
requests==2.31.0
Pillow>=10.1.0
protobuf==3.20.3
grpcio==1.53.0
grpcio-tools==1.53.0
stability-sdk==0.8.3 